+++ src/Core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## NOT RELEASED
+
+### Added
+
+- Support for EventBridge in `AwsClientFactory`
+
 ## 1.1.0
 
 ### Added